third_party/blink/web_tests/http/tests/preload/warning/integrity-mismatch.html

<!doctype html>
<html>
<link rel="preload" as="script" href="resources/script.js"></link>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<body>
<script src="resources/script.js" integrity="sha384-qfsBakWU2bdKoSdjyv97lIxUQH9+clPjBB/8ExlmEdvB6R/XCard/w4M5y82DA8D"></script>
<script>
promise_test(() => {
  return new Promise(resolve => {
    window.addEventListener('load', resolve);
  });
}, "Preload match fails due to an integrity mismatch (see the console log).");
</script>
</body>
</html>
