third_party/blink/web_tests/http/tests/preload/warning/integrity-mismatch-expected.txt

CONSOLE WARNING: A preload for 'http://127.0.0.1:8000/preload/warning/resources/script.js' is found, but is not used due to an integrity mismatch.
This is a testharness.js-based test.
Harness: the test ran to completion.
